Creating repository for my undergraduate honors senior thesis project, segmenting CT scans of intracranial hemorrhage with a 3D U-net.
The main code to run the network is in 3d_model.py, with helping files in the model directory.
The ICH image data should be stored in directories named npy-img and npy-mask, under the main directory. These files contain cleaned and normalized data for the ICH CT scans, stored in .npy format.
The CT Scans are from the dataset found at https://physionet.org/content/ct-ich/1.3.1/.